game loop is done
